ebisu-bench: basic implementation
This is a basic implementation of the import idea, however some quick
tests have found that Ebisu has some pretty major flaws in that it
doesn't actually make use of the more modern model of memory (it only
makes use of the forgetting curve but not the spacing effect) -- meaning
that its internal model is always chasing after the real recall curve.

This is something they are planning on fixing[1] but we should look into
alternative algorithms.

[1]: https://github.com/fasiha/ebisu/issues/35

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>